+++ CHANGELOG.md
@@ -33,6 +33,7 @@
             - Config entries with a value of `undefined` are ignored
         - Added a default image-like content string detection annotation in the `struct` view: when a string value is identified as image-like content, a badge is displayed before the string, showing an image preview on hover
 - Added support for `diff` syntax in `source` view
+- Added `hideOnTriggerClick` option for popup and tooltip views
 - Reworked the computation graph on the discovery page and related features:
     - Enhanced the look and feel of the graph
     - Improved layout handling for complex graphs